[runtime] Enhancment: Move OperationResult out of inetstack
