+++ src/rust/demikernel/libos/network.rs
@@ -33,12 +33,6 @@ use crate::catnip::CatnipLibOS;
 use crate::catpowder::CatpowderLibOS;
 
 //======================================================================================================================
-// Exports
-//======================================================================================================================
-
-pub use crate::inetstack::operations::OperationResult;
-
-//======================================================================================================================
 // Structures
 //======================================================================================================================
