+++ src/rust/runtime/queue/qresult.rs
@@ -7,14 +7,29 @@
 
 use crate::runtime::{
     fail::Fail,
+    memory::DemiBuffer,
     network::types::Port16,
     QDesc,
 };
-use ::std::net::Ipv4Addr;
+use ::std::{
+    fmt,
+    net::{
+        Ipv4Addr,
+        SocketAddrV4,
+    },
+};
 
 //==============================================================================
-// Enumerations
+// Structures
 //==============================================================================
+
+pub enum OperationResult {
+    Connect,
+    Accept((QDesc, SocketAddrV4)),
+    Push,
+    Pop(Option<SocketAddrV4>, DemiBuffer),
+    Failed(Fail),
+}
 
 /// Result for IO Queue Operations
 pub enum QResult {
@@ -24,4 +39,19 @@ pub enum QResult {
     PushTo,
     Pop(Option<(Ipv4Addr, Port16)>, Vec<u8>),
     Failed(Fail),
+}
+//==============================================================================
+// Trait Implementations
+//==============================================================================
+
+impl fmt::Debug for OperationResult {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        match self {
+            OperationResult::Connect => write!(f, "Connect"),
+            OperationResult::Accept(..) => write!(f, "Accept"),
+            OperationResult::Push => write!(f, "Push"),
+            OperationResult::Pop(..) => write!(f, "Pop"),
+            OperationResult::Failed(ref e) => write!(f, "Failed({:?})", e),
+        }
+    }
 }
